README updated. Unused config removed from app/configs/app_config.py Unused string removed and new language list added to app/configs/config.json. app/database/ folder moved to app/services/. app/services/sum_methods.py file renamed to sum_file_methods.py. Now the value in app/static/index.html is not languages, but only language shortcodes.
The app/routers/youtube.py file has been improved using the above changes.
Now the names of summary files have the summary language at the end of the name (in case of translation).
A function has been added to check for an existing transcript, and if successful, the existing transcript is returned to the user without re-summarization (and translation).